+++ Sharp3D.Math/Core/Vector2D.cs
@@ -95,8 +95,8 @@ namespace Sharp3D.Math.Core
 		/// <param name="context">The contextual information about the source or destination.</param>
 		private Vector2D(SerializationInfo info, StreamingContext context)
 		{
-			_x = info.GetSingle("X");
-			_y = info.GetSingle("Y");
+			_x = info.GetDouble("X");
+            _y = info.GetDouble("Y");
 		}
 		#endregion
