+++ Sharp3D.Math/Core/Vector4D.cs
@@ -107,10 +107,10 @@ namespace Sharp3D.Math.Core
 		/// <param name="context">The contextual information about the source or destination.</param>
 		private Vector4D(SerializationInfo info, StreamingContext context)
 		{
-			_x = info.GetSingle("X");
-			_y = info.GetSingle("Y");
-			_z = info.GetSingle("Z");
-			_w = info.GetSingle("W");
+            _x = info.GetDouble("X");
+            _y = info.GetDouble("Y");
+            _z = info.GetDouble("Z");
+            _w = info.GetDouble("W");
 		}
 		#endregion
